Remove toggle and move files
