+++ pkg/services/featuremgmt/toggles_gen.csv
@@ -87,7 +87,6 @@ exploreScrollableLogsContainer,experimental,@grafana/observability-logs,false,fa
 recordedQueriesMulti,experimental,@grafana/observability-metrics,false,false,false,false
 pluginsDynamicAngularDetectionPatterns,experimental,@grafana/plugins-platform-backend,false,false,false,false
 alertingLokiRangeToInstant,experimental,@grafana/alerting-squad,false,false,false,false
-flameGraphV2,experimental,@grafana/observability-traces-and-profiling,false,false,false,true
 elasticToggleableFilters,experimental,@grafana/observability-logs,false,false,false,true
 vizAndWidgetSplit,experimental,@grafana/dashboards-squad,false,false,false,true
 prometheusIncrementalQueryInstrumentation,experimental,@grafana/observability-metrics,false,false,false,true
