+++ public/app/plugins/panel/flamegraph/components/FlameGraph/FlameGraph.test.tsx
@@ -1,12 +1,12 @@
 import { fireEvent, render, screen } from '@testing-library/react';
-import React, { useState } from 'react';
+import React from 'react';
 
 import { createDataFrame } from '@grafana/data';
 
-import { SelectedView } from '../types';
+import { ColorScheme } from '../types';
 
 import FlameGraph from './FlameGraph';
-import { FlameGraphDataContainer, nestedSetToLevels } from './dataTransform';
+import { FlameGraphDataContainer } from './dataTransform';
 import { data } from './testData/dataNestedSet';
 
 import 'jest-canvas-mock';
@@ -23,42 +23,52 @@ jest.mock('react-use', () => {
 });
 
 describe('FlameGraph', () => {
-  const FlameGraphWithProps = () => {
-    const [focusedItemIndex, setFocusedItemIndex] = useState<number>();
-    const [rangeMin, setRangeMin] = useState(0);
-    const [rangeMax, setRangeMax] = useState(1);
-    const [search] = useState('');
-    const [selectedView, _] = useState(SelectedView.Both);
-
+  function setup() {
     const flameGraphData = createDataFrame(data);
     const container = new FlameGraphDataContainer(flameGraphData);
-    const levels = nestedSetToLevels(container);
 
-    return (
+    const setRangeMin = jest.fn();
+    const setRangeMax = jest.fn();
+    const onItemFocused = jest.fn();
+    const onSandwich = jest.fn();
+    const onFocusPillClick = jest.fn();
+    const onSandwichPillClick = jest.fn();
+
+    const renderResult = render(
       <FlameGraph
         data={container}
-        levels={levels}
-        rangeMin={rangeMin}
-        rangeMax={rangeMax}
-        search={search}
+        rangeMin={0}
+        rangeMax={1}
+        search={''}
         setRangeMin={setRangeMin}
         setRangeMax={setRangeMax}
-        selectedView={selectedView}
-        onItemFocused={(itemIndex) => {
-          setFocusedItemIndex(itemIndex);
-        }}
-        focusedItemIndex={focusedItemIndex}
+        onItemFocused={onItemFocused}
         textAlign={'left'}
+        onSandwich={onSandwich}
+        onFocusPillClick={onFocusPillClick}
+        onSandwichPillClick={onSandwichPillClick}
+        colorScheme={ColorScheme.ValueBased}
       />
     );
-  };
+    return {
+      renderResult,
+      mocks: {
+        setRangeMax,
+        setRangeMin,
+        onItemFocused,
+        onSandwich,
+        onFocusPillClick,
+        onSandwichPillClick,
+      },
+    };
+  }
 
   it('should render without error', async () => {
-    expect(() => render(<FlameGraphWithProps />)).not.toThrow();
+    setup();
   });
 
   it('should render correctly', async () => {
-    render(<FlameGraphWithProps />);
+    setup();
 
     const canvas = screen.getByTestId('flameGraph') as HTMLCanvasElement;
     const ctx = canvas!.getContext('2d');
@@ -67,20 +77,21 @@ describe('FlameGraph', () => {
   });
 
   it('should render metadata', async () => {
-    render(<FlameGraphWithProps />);
-    expect(screen.getByText('16.5 Bil (100%) of 16,460,000,000 total samples (Count)')).toBeDefined();
+    setup();
+    expect(screen.getByText('16.5 Bil | 16.5 Bil samples (Count)')).toBeDefined();
   });
 
   it('should render context menu', async () => {
-    const event = new MouseEvent('click');
+    const event = new MouseEvent('click', { bubbles: true });
     Object.defineProperty(event, 'offsetX', { get: () => 10 });
     Object.defineProperty(event, 'offsetY', { get: () => 10 });
     Object.defineProperty(HTMLCanvasElement.prototype, 'clientWidth', { configurable: true, value: 500 });
 
-    const screen = render(<FlameGraphWithProps />);
+    setup();
     const canvas = screen.getByTestId('flameGraph') as HTMLCanvasElement;
     expect(canvas).toBeInTheDocument();
     expect(screen.queryByTestId('contextMenu')).not.toBeInTheDocument();
+
     fireEvent(canvas, event);
     expect(screen.getByTestId('contextMenu')).toBeInTheDocument();
   });
