+++ public/app/plugins/panel/flamegraph/components/FlameGraph/colors.ts
