+++ public/app/plugins/panel/flamegraph/components/FlameGraph/rendering.ts
@@ -1,4 +1,8 @@
-import { colors } from '@grafana/ui';
+import uFuzzy from '@leeoniya/ufuzzy';
+import { RefObject, useEffect, useMemo, useState } from 'react';
+
+import { GrafanaTheme2 } from '@grafana/data';
+import { colors, useTheme2 } from '@grafana/ui';
 
 import {
   BAR_BORDER_WIDTH,
@@ -8,9 +12,115 @@ import {
   LABEL_THRESHOLD,
   PIXELS_PER_LEVEL,
 } from '../../constants';
-import { TextAlign } from '../types';
+import { ClickedItemData, ColorScheme, TextAlign } from '../types';
 
+import { getBarColorByPackage, getBarColorByValue } from './colors';
 import { FlameGraphDataContainer, LevelItem } from './dataTransform';
+
+const ufuzzy = new uFuzzy();
+
+export function useFlameRender(
+  canvasRef: RefObject<HTMLCanvasElement>,
+  data: FlameGraphDataContainer,
+  levels: LevelItem[][],
+  wrapperWidth: number,
+  rangeMin: number,
+  rangeMax: number,
+  search: string,
+  textAlign: TextAlign,
+  totalTicks: number,
+  colorScheme: ColorScheme,
+  focusedItemData?: ClickedItemData
+) {
+  const foundLabels = useMemo(() => {
+    if (search) {
+      const foundLabels = new Set<string>();
+      let idxs = ufuzzy.filter(data.getUniqueLabels(), search);
+
+      if (idxs) {
+        for (let idx of idxs) {
+          foundLabels.add(data.getUniqueLabels()[idx]);
+        }
+      }
+
+      return foundLabels;
+    }
+    // In this case undefined means there was no search so no attempt to highlighting anything should be made.
+    return undefined;
+  }, [search, data]);
+
+  const ctx = useSetupCanvas(canvasRef, wrapperWidth, levels.length);
+  const theme = useTheme2();
+
+  useEffect(() => {
+    if (!ctx) {
+      return;
+    }
+    ctx.clearRect(0, 0, ctx.canvas.width, ctx.canvas.height);
+    const pixelsPerTick = (wrapperWidth * window.devicePixelRatio) / totalTicks / (rangeMax - rangeMin);
+
+    for (let levelIndex = 0; levelIndex < levels.length; levelIndex++) {
+      const level = levels[levelIndex];
+      // Get all the dimensions of the rectangles for the level. We do this by level instead of per rectangle, because
+      // sometimes we collapse multiple bars into single rect.
+      const dimensions = getRectDimensionsForLevel(data, level, levelIndex, totalTicks, rangeMin, pixelsPerTick);
+      for (const rect of dimensions) {
+        const focusedLevel = focusedItemData ? focusedItemData.level : 0;
+        // Render each rectangle based on the computed dimensions
+        renderRect(
+          ctx,
+          rect,
+          totalTicks,
+          rangeMin,
+          rangeMax,
+          levelIndex,
+          focusedLevel,
+          foundLabels,
+          textAlign,
+          colorScheme,
+          theme
+        );
+      }
+    }
+  }, [
+    ctx,
+    data,
+    levels,
+    wrapperWidth,
+    rangeMin,
+    rangeMax,
+    search,
+    focusedItemData,
+    foundLabels,
+    textAlign,
+    totalTicks,
+    colorScheme,
+    theme,
+  ]);
+}
+
+function useSetupCanvas(canvasRef: RefObject<HTMLCanvasElement>, wrapperWidth: number, numberOfLevels: number) {
+  const [ctx, setCtx] = useState<CanvasRenderingContext2D>();
+
+  useEffect(() => {
+    if (!(numberOfLevels && canvasRef.current)) {
+      return;
+    }
+    const ctx = canvasRef.current.getContext('2d')!;
+
+    const height = PIXELS_PER_LEVEL * numberOfLevels;
+    canvasRef.current.width = Math.round(wrapperWidth * window.devicePixelRatio);
+    canvasRef.current.height = Math.round(height);
+    canvasRef.current.style.width = `${wrapperWidth}px`;
+    canvasRef.current.style.height = `${height / window.devicePixelRatio}px`;
+
+    ctx.textBaseline = 'middle';
+    ctx.font = 12 * window.devicePixelRatio + 'px monospace';
+    ctx.strokeStyle = 'white';
+    setCtx(ctx);
+  }, [canvasRef, setCtx, wrapperWidth, numberOfLevels]);
+  return ctx;
+}
 
 type RectData = {
   width: number;
@@ -40,7 +150,7 @@ export function getRectDimensionsForLevel(
   for (let barIndex = 0; barIndex < level.length; barIndex += 1) {
     const item = level[barIndex];
     const barX = getBarX(item.start, totalTicks, rangeMin, pixelsPerTick);
-    let curBarTicks = data.getValue(item.itemIndex);
+    let curBarTicks = item.value;
 
     // merge very small blocks into big "collapsed" ones for performance
     const collapsed = curBarTicks * pixelsPerTick <= COLLAPSE_THRESHOLD;
@@ -48,14 +158,14 @@ export function getRectDimensionsForLevel(
       while (
         barIndex < level.length - 1 &&
         item.start + curBarTicks === level[barIndex + 1].start &&
-        data.getValue(level[barIndex + 1].itemIndex) * pixelsPerTick <= COLLAPSE_THRESHOLD
+        level[barIndex + 1].value * pixelsPerTick <= COLLAPSE_THRESHOLD
       ) {
         barIndex += 1;
-        curBarTicks += data.getValue(level[barIndex].itemIndex);
+        curBarTicks += level[barIndex].value;
       }
     }
 
-    const displayValue = data.getValueDisplay(item.itemIndex);
+    const displayValue = data.valueDisplayProcessor(item.value);
     let unit = displayValue.suffix ? displayValue.text + displayValue.suffix : displayValue.text;
 
     const width = curBarTicks * pixelsPerTick - (collapsed ? 0 : BAR_BORDER_WIDTH * 2);
@@ -66,9 +176,9 @@ export function getRectDimensionsForLevel(
       y: levelIndex * PIXELS_PER_LEVEL,
       collapsed,
       ticks: curBarTicks,
-      label: data.getLabel(item.itemIndex),
+      label: data.getLabel(item.itemIndexes[0]),
       unitLabel: unit,
-      itemIndex: item.itemIndex,
+      itemIndex: item.itemIndexes[0],
     });
   }
   return coordinatesLevel;
@@ -80,11 +190,12 @@ export function renderRect(
   totalTicks: number,
   rangeMin: number,
   rangeMax: number,
-  query: string,
   levelIndex: number,
   topLevelIndex: number,
-  foundNames: Set<string>,
-  textAlign: TextAlign
+  foundNames: Set<string> | undefined,
+  textAlign: TextAlign,
+  colorScheme: ColorScheme,
+  theme: GrafanaTheme2
 ) {
   if (rect.width < HIDE_THRESHOLD) {
     return;
@@ -93,28 +204,36 @@ export function renderRect(
   ctx.beginPath();
   ctx.rect(rect.x + (rect.collapsed ? 0 : BAR_BORDER_WIDTH), rect.y, rect.width, rect.height);
 
-  //  / (rangeMax - rangeMin) here so when you click a bar it will adjust the top (clicked)bar to the most 'intense' color
-  const intensity = Math.min(1, rect.ticks / totalTicks / (rangeMax - rangeMin));
-  const h = 50 - 50 * intensity;
-  const l = 65 + 7 * intensity;
+  const color =
+    colorScheme === ColorScheme.ValueBased
+      ? getBarColorByValue(rect.ticks, totalTicks, rangeMin, rangeMax)
+      : getBarColorByPackage(rect.label, theme);
 
-  const name = rect.label;
-
-  if (!rect.collapsed) {
-    ctx.stroke();
-
-    if (query) {
-      ctx.fillStyle = foundNames.has(name) ? getBarColor(h, l) : colors[55];
-    } else {
-      ctx.fillStyle = levelIndex > topLevelIndex - 1 ? getBarColor(h, l) : getBarColor(h, l + 15);
-    }
+  if (foundNames) {
+    // Means we are searching, we use color for matches and gray the rest
+    ctx.fillStyle = foundNames.has(rect.label) ? color.toHslString() : colors[55];
   } else {
-    ctx.fillStyle = foundNames.has(name) ? getBarColor(h, l) : colors[55];
+    // No search
+    if (rect.collapsed) {
+      // Collapsed are always grayed
+      ctx.fillStyle = colors[55];
+    } else {
+      // Mute if we are above the focused symbol
+      ctx.fillStyle = levelIndex > topLevelIndex - 1 ? color.toHslString() : color.lighten(15).toHslString();
+    }
   }
+
+  if (rect.collapsed) {
+    // Only fill the collapsed rects
+    ctx.fill();
+    return;
+  }
+
+  ctx.stroke();
   ctx.fill();
 
-  if (!rect.collapsed && rect.width >= LABEL_THRESHOLD) {
-    renderLabel(ctx, name, rect, textAlign);
+  if (rect.width >= LABEL_THRESHOLD) {
+    renderLabel(ctx, rect.label, rect, textAlign);
   }
 }
 
@@ -158,8 +277,4 @@ function renderLabel(ctx: CanvasRenderingContext2D, name: string, rect: RectData
  */
 export function getBarX(offset: number, totalTicks: number, rangeMin: number, pixelsPerTick: number) {
   return (offset - totalTicks * rangeMin) * pixelsPerTick;
-}
-
-function getBarColor(h: number, l: number) {
-  return `hsl(${h}, 100%, ${l}%)`;
 }
