+++ public/app/plugins/panel/flamegraph/components/FlameGraphContainer.tsx
@@ -9,10 +9,10 @@ import { useStyles2, useTheme2 } from '@grafana/ui';
 import { MIN_WIDTH_TO_SHOW_BOTH_TOPTABLE_AND_FLAMEGRAPH } from '../constants';
 
 import FlameGraph from './FlameGraph/FlameGraph';
-import { FlameGraphDataContainer, LevelItem, nestedSetToLevels } from './FlameGraph/dataTransform';
+import { FlameGraphDataContainer } from './FlameGraph/dataTransform';
 import FlameGraphHeader from './FlameGraphHeader';
 import FlameGraphTopTableContainer from './TopTable/FlameGraphTopTableContainer';
-import { SelectedView, TextAlign } from './types';
+import { ClickedItemData, ColorScheme, SelectedView, TextAlign } from './types';
 
 type Props = {
   data?: DataFrame;
@@ -20,7 +20,7 @@ type Props = {
 };
 
 const FlameGraphContainer = (props: Props) => {
-  const [focusedItemIndex, setFocusedItemIndex] = useState<number>();
+  const [focusedItemData, setFocusedItemData] = useState<ClickedItemData>();
 
   const [rangeMin, setRangeMin] = useState(0);
   const [rangeMax, setRangeMax] = useState(1);
@@ -28,19 +28,17 @@ const FlameGraphContainer = (props: Props) => {
   const [selectedView, setSelectedView] = useState(SelectedView.Both);
   const [sizeRef, { width: containerWidth }] = useMeasure<HTMLDivElement>();
   const [textAlign, setTextAlign] = useState<TextAlign>('left');
+  // This is a label of the item because in sandwich view we group all items by label and present a merged graph
+  const [sandwichItem, setSandwichItem] = useState<string>();
+  const [colorScheme, setColorScheme] = useState<ColorScheme>(ColorScheme.ValueBased);
 
   const theme = useTheme2();
 
-  const [dataContainer, levels] = useMemo((): [FlameGraphDataContainer, LevelItem[][]] | [undefined, undefined] => {
+  const dataContainer = useMemo((): FlameGraphDataContainer | undefined => {
     if (!props.data) {
-      return [undefined, undefined];
+      return;
     }
-    const container = new FlameGraphDataContainer(props.data, theme);
-
-    // Transform dataFrame with nested set format to array of levels. Each level contains all the bars for a particular
-    // level of the flame graph. We do this temporary as in the end we should be able to render directly by iterating
-    // over the dataFrame rows.
-    return [container, nestedSetToLevels(container)];
+    return new FlameGraphDataContainer(props.data, theme);
   }, [props.data, theme]);
 
   const styles = useStyles2(getStyles);
@@ -56,10 +54,19 @@ const FlameGraphContainer = (props: Props) => {
     }
   }, [selectedView, setSelectedView, containerWidth]);
 
-  useEffect(() => {
-    setFocusedItemIndex(undefined);
+  function resetFocus() {
+    setFocusedItemData(undefined);
     setRangeMin(0);
     setRangeMax(1);
+  }
+
+  function resetSandwich() {
+    setSandwichItem(undefined);
+  }
+
+  useEffect(() => {
+    resetFocus();
+    resetSandwich();
   }, [props.data]);
 
   return (
@@ -74,12 +81,14 @@ const FlameGraphContainer = (props: Props) => {
             setSelectedView={setSelectedView}
             containerWidth={containerWidth}
             onReset={() => {
-              setRangeMin(0);
-              setRangeMax(1);
-              setFocusedItemIndex(undefined);
+              resetFocus();
+              resetSandwich();
             }}
             textAlign={textAlign}
             onTextAlignChange={setTextAlign}
+            showResetButton={Boolean(focusedItemData || sandwichItem)}
+            colorScheme={colorScheme}
+            onColorSchemeChange={setColorScheme}
           />
 
           <div className={styles.body}>
@@ -87,7 +96,6 @@ const FlameGraphContainer = (props: Props) => {
               <FlameGraphTopTableContainer
                 data={dataContainer}
                 app={props.app}
-                totalLevels={levels.length}
                 onSymbolClick={(symbol) => {
                   if (search === symbol) {
                     setSearch('');
@@ -97,27 +105,31 @@ const FlameGraphContainer = (props: Props) => {
                       grafana_version: config.buildInfo.version,
                     });
                     setSearch(symbol);
-                    // Reset selected level in flamegraph when selecting row in top table
-                    setRangeMin(0);
-                    setRangeMax(1);
                   }
                 }}
+                height={selectedView === SelectedView.TopTable ? 600 : undefined}
               />
             )}
 
             {selectedView !== SelectedView.TopTable && (
               <FlameGraph
                 data={dataContainer}
-                levels={levels}
                 rangeMin={rangeMin}
                 rangeMax={rangeMax}
                 search={search}
                 setRangeMin={setRangeMin}
                 setRangeMax={setRangeMax}
-                selectedView={selectedView}
-                onItemFocused={(itemIndex) => setFocusedItemIndex(itemIndex)}
-                focusedItemIndex={focusedItemIndex}
+                onItemFocused={(data) => setFocusedItemData(data)}
+                focusedItemData={focusedItemData}
                 textAlign={textAlign}
+                sandwichItem={sandwichItem}
+                onSandwich={(label: string) => {
+                  resetFocus();
+                  setSandwichItem(label);
+                }}
+                onFocusPillClick={resetFocus}
+                onSandwichPillClick={resetSandwich}
+                colorScheme={colorScheme}
               />
             )}
           </div>
