--- public/app/plugins/panel/flamegraph/flamegraphV2/constants.ts
@@ -1,8 +1,0 @@
-export const PIXELS_PER_LEVEL = 22 * window.devicePixelRatio;
-export const COLLAPSE_THRESHOLD = 10 * window.devicePixelRatio;
-export const HIDE_THRESHOLD = 0.5 * window.devicePixelRatio;
-export const LABEL_THRESHOLD = 20 * window.devicePixelRatio;
-export const BAR_BORDER_WIDTH = 0.5 * window.devicePixelRatio;
-export const BAR_TEXT_PADDING_LEFT = 4 * window.devicePixelRatio;
-export const MIN_WIDTH_TO_SHOW_BOTH_TOPTABLE_AND_FLAMEGRAPH = 800;
-export const TOP_TABLE_COLUMN_WIDTH = 120;
