+++ public/app/plugins/panel/flamegraph/suggestions.ts
@@ -1,9 +1,7 @@
 import { VisualizationSuggestionsBuilder } from '@grafana/data';
-import { config } from '@grafana/runtime';
 import { SuggestionName } from 'app/types/suggestions';
 
 import { FlameGraphDataContainer as FlameGraphDataContainer } from './components/FlameGraph/dataTransform';
-import { FlameGraphDataContainer as FlameGraphDataContainerV2 } from './flamegraphV2/components/FlameGraph/dataTransform';
 
 export class FlameGraphSuggestionsSupplier {
   getListWithDefaults(builder: VisualizationSuggestionsBuilder) {
@@ -23,9 +21,7 @@ export class FlameGraphSuggestionsSupplier {
     // Without this check, a suggestion containing an error is shown to the user.
     const dataFrame = builder.data.series[0];
     try {
-      config.featureToggles.flameGraphV2
-        ? new FlameGraphDataContainerV2(dataFrame)
-        : new FlameGraphDataContainer(dataFrame);
+      new FlameGraphDataContainer(dataFrame);
     } catch (err) {
       return;
     }
